requirements.txt
alembic==1.3.2
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
config==0.4.2
dominate==2.4.0
environs==7.1.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
Pillow==6.2.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
